simplify Network.JotForm.Core
On reflection, the design of `toRequest` was not ideal. It will be
better to simply take in the query string and request body as
arguments, and handle any extra logic about special cases for
different methods separately.

As part of this change, we bundle up a couple parameters into a
`Params` datatype, with a default constructor for ease of use.

In Network.JotForm.Api we refactor the body of the existing methods
to use `Params`, and explicitly pass extra headers when needed.